+++ strnaming/classes.py
@@ -1,5 +1,5 @@
 #
-# Copyright (C) 2021 Jerry Hoogenboom
+# Copyright (C) 2023 Jerry Hoogenboom
 #
 # This file is part of STRNaming, an algorithm for generating simple,
 # informative names for sequenced STR alleles in a standardised and
@@ -132,7 +132,18 @@ class ReferenceStructureStore:
 
         The end position is exclusive.
         """
-        for structure in reference_structures.get_within_range(chromosome, start, end - 1):
+        for structure in reference_structures.gen_within_range(chromosome, start, end - 1):
+            self.add_structure(chromosome, structure)
+
+    def load_within_ranges(self, chromosome, ranges):
+        """
+        Load human genome reference STR structure data.
+
+        The ranges should be a sorted iterable of (start, end) pairs.
+        The end positions are exclusive.
+        """
+        for structure in reference_structures.gen_within_any_range(
+                chromosome, ((start, end - 1) for start, end in ranges)):
             self.add_structure(chromosome, structure)
 
     def load_from_tsv(self, input):